+++ src/Repositories/MainRepository.js
@@ -4,10 +4,9 @@ const { capitalCase, snakeCase } = require('change-case')
 const HttpException = require('./../Exceptions/HttpException')
 
 class MainRepository {
-  constructor (validation, route, trigger, repositoryHelper, queryParser, event) {
+  constructor (validation, route, repositoryHelper, queryParser, event) {
     this.validation = validation
     this.route = route
-    this.trigger = trigger
     this.repositoryHelper = repositoryHelper
     this.queryParser = queryParser
     this.event = event
@@ -35,9 +34,15 @@ class MainRepository {
     // Users should be able to add relationships to the query
     this.queryParser.applyRelations(query, conditions.with)
 
-    // We should trigger onBeforePagination events
+    // We should call onBeforePagination action
     const modelName = modelPath.replace('App/Models/', '')
-    await this.trigger.fire(`onBeforePaginate${modelName}`, { request, params, query })
+
+    await this.repositoryHelper.callAction(
+      request.adonisx.url,
+      'onBeforePaginate',
+      { request, params, query }
+    )
+
     this.event.fire(`onBeforePaginate${modelName}`, { request, params, query })
 
     // User should be able to select sorting fields and types
@@ -49,8 +54,13 @@ class MainRepository {
       conditions.per_page
     )
 
-    // We should trigger onAfterPagination events
-    await this.trigger.fire(`onAfterPaginate${modelName}`, { request, params, result })
+    // We should call onAfterPagination action
+    await this.repositoryHelper.callAction(
+      request.adonisx.url,
+      'onAfterPaginate',
+      { request, params, result }
+    )
+
     this.event.fire(`onAfterPaginate${modelName}`, { request, params, result })
 
     // And this is my function response
@@ -82,17 +92,28 @@ class MainRepository {
     // We should add this condition in here because of performance.
     query.where('id', params.id)
 
-    // We should trigger onBeforeFirstOrFail events
+    // We should call onBeforeShow action
     const modelName = modelPath.replace('App/Models/', '')
-    await this.trigger.fire(`onBeforeShow${modelName}`, { request, params, query })
+
+    await this.repositoryHelper.callAction(
+      request.adonisx.url,
+      'onBeforeShow',
+      { request, params, query }
+    )
+
     this.event.fire(`onBeforeShow${modelName}`, { request, params, query })
 
     const item = await this.safe(request, async () => {
       return await query.firstOrFail()
     })
 
-    // We should trigger onAfterFirstOrFail events
-    await this.trigger.fire(`onAfterShow${modelName}`, { request, params, item })
+    // We should call onAfterShow action
+    await this.repositoryHelper.callAction(
+      request.adonisx.url,
+      'onAfterShow',
+      { request, params, item }
+    )
+
     this.event.fire(`onAfterShow${modelName}`, { request, params, item })
 
     return item
@@ -113,16 +134,27 @@ class MainRepository {
       data[snakeCase(request.adonisx.parent_column)] = params[request.adonisx.parent_column]
     }
 
-    // We should trigger onBeforeCreate events
+    // We should call onBeforeCreate action
     const modelName = modelPath.replace('App/Models/', '')
-    await this.trigger.fire(`onBeforeCreate${modelName}`, { request, params, data })
+
+    await this.repositoryHelper.callAction(
+      request.adonisx.url,
+      'onBeforeCreate',
+      { request, params, data }
+    )
+
     this.event.fire(`onBeforeCreate${modelName}`, { request, params, data })
 
     // Creating the item
     const item = await Model.create(data)
 
-    // We should trigger onAfterCreate events
-    await this.trigger.fire(`onAfterCreate${modelName}`, { request, params, data, item })
+    // We should call onAfterCreate action
+    await this.repositoryHelper.callAction(
+      request.adonisx.url,
+      'onAfterCreate',
+      { request, params, data, item }
+    )
+
     this.event.fire(`onAfterCreate${modelName}`, { request, params, data, item })
 
     // Returning response
@@ -139,14 +171,23 @@ class MainRepository {
 
     this.repositoryHelper.addParentIdCondition(query, params, request.adonisx.parent_column)
 
-    // We should trigger onBeforeUpdateQuery events
+    // We should call onBeforeUpdateQuery action
     const modelName = modelPath.replace('App/Models/', '')
-    await this.trigger.fire(`onBeforeUpdateQuery${modelName}`, { request, params, query })
+
+    await this.repositoryHelper.callAction(
+      request.adonisx.url,
+      'onBeforeUpdateQuery',
+      { request, params, query }
+    )
 
     const item = await query.firstOrFail()
 
-    // We should trigger onAfterUpdateQuery events
-    await this.trigger.fire(`onAfterUpdateQuery${modelName}`, { request, params, item })
+    // We should call onAfterUpdateQuery action
+    await this.repositoryHelper.callAction(
+      request.adonisx.url,
+      'onAfterUpdateQuery',
+      { request, params, item }
+    )
 
     // Preparing the data
     const data = request.getFillableFields(Model)
@@ -162,14 +203,22 @@ class MainRepository {
     // We should validate the data
     await this.validation.validate(request.method(), item.toJSON(), Model.validations)
 
-    // We should trigger onBeforeUpdate events
-    await this.trigger.fire(`onBeforeUpdate${modelName}`, { request, params, item })
+    // We should call onBeforeUpdate action
+    await this.repositoryHelper.callAction(
+      request.adonisx.url,
+      'onBeforeUpdate',
+      { request, params, item }
+    )
     this.event.fire(`onBeforeUpdate${modelName}`, { request, params, item })
 
     await item.save()
 
-    // We should trigger onAfterUpdate events
-    await this.trigger.fire(`onAfterUpdate${modelName}`, { request, params, item })
+    // We should call onAfterUpdate action
+    await this.repositoryHelper.callAction(
+      request.adonisx.url,
+      'onAfterUpdate',
+      { request, params, item }
+    )
     this.event.fire(`onAfterUpdate${modelName}`, { request, params, item })
 
     // Returning response
@@ -189,9 +238,14 @@ class MainRepository {
     // Appending parent id condition
     this.repositoryHelper.addParentIdCondition(query, params, request.adonisx.parent_column)
 
-    // We should trigger onBeforeDelete events
+    // We should call onBeforeDelete action
     const modelName = modelPath.replace('App/Models/', '')
-    await this.trigger.fire(`onBeforeDelete${modelName}`, { request, params, query })
+
+    await this.repositoryHelper.callAction(
+      request.adonisx.url,
+      'onBeforeDelete',
+      { request, params, query }
+    )
     this.event.fire(`onBeforeDelete${modelName}`, { request, params, query })
 
     const item = (await query.firstOrFail()).toJSON()
@@ -199,8 +253,13 @@ class MainRepository {
     // Deleting the item
     await query.delete()
 
-    // We should trigger onAfterDelete events
-    await this.trigger.fire(`onAfterDelete${modelName}`, { request, params, item })
+    // We should call onAfterDelete action
+    await this.repositoryHelper.callAction(
+      request.adonisx.url,
+      'onAfterDelete',
+      { request, params, item }
+    )
+
     this.event.fire(`onAfterDelete${modelName}`, { request, params, item })
   }