+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.3.0
+
+- Trigger feature has been renamed as **Actions** and auto definition structure has been created. ([#8](https://github.com/adonisx/adonisx/issues/8))
+- Method base fillable field feature and method based validations have been implemented. ([#7](https://github.com/adonisx/adonisx/issues/7))
+
 ## 0.2.2
 
 - Fixed Like queries [#2](https://github.com/adonisx/adonisx/issues/2)